+++ inst/examples/oauth_module_server.R
@@ -43,9 +43,13 @@ if (
 
   # Example app with auto-redirect (1) -----------------------------------------
 
-  ui_1 <- oauth_ui(fluidPage(
-    uiOutput("login")
-  ))
+  ui_1 <- oauth_ui(
+    fluidPage(
+      uiOutput("login")
+    ),
+    id = "auth",
+    client = client
+  )
 
   server_1 <- function(input, output, session) {
     # Auto-redirect (default):
@@ -79,11 +83,15 @@ if (
 
   # Example app with manual login button (2) -----------------------------------
 
-  ui_2 <- oauth_ui(fluidPage(
-    actionButton("login_btn", "Login"),
-    actionButton("logout_btn", "Logout"),
-    uiOutput("login")
-  ))
+  ui_2 <- oauth_ui(
+    fluidPage(
+      actionButton("login_btn", "Login"),
+      actionButton("logout_btn", "Logout"),
+      uiOutput("login")
+    ),
+    id = "auth",
+    client = client
+  )
 
   server_2 <- function(input, output, session) {
     auth <- oauth_module_server(
@@ -126,9 +134,13 @@ if (
   # Below app shows the authenticated username + their GitHub repositories,
   # fetched via GitHub API using the access token obtained during login
 
-  ui_3 <- oauth_ui(fluidPage(
-    uiOutput("ui")
-  ))
+  ui_3 <- oauth_ui(
+    fluidPage(
+      uiOutput("ui")
+    ),
+    id = "auth",
+    client = client
+  )
 
   server_3 <- function(input, output, session) {
     auth <- oauth_module_server(